add search bar and photo at home page
