+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Routes, Route } from 'react-router-dom'
-import Home from './pages/home.jsx'
+import Home from './pages/home/home.jsx'
 
 function App() {
   return (
